Move the function with fs calls to another module
