+++ packages/design-system-tokens/src/lib/tokens.ts
@@ -1,6 +1,3 @@
-import * as fs from 'fs';
-import * as path from 'path';
-
 import { VariableCodeSyntax, VariableScope } from '@figma/rest-api-spec';
 
 /**
@@ -71,11 +68,11 @@ export type FlattenedTokensByFile = {
  * object, producing a key that contains each group and property in dot notation (e.g.,
  * `color.crimson.500`). This makes each token easier to reference and transform.
  */
-export function flattenTokensFile(tokensFile: TokensFile) {
+export function flattenTokens(tokensFile: TokensFile) {
   const flattenedTokens: { [tokenName: string]: Token } = {};
 
   Object.entries(tokensFile).forEach(([tokenGroup, groupValues]) => {
-    traverseCollection({ key: tokenGroup, object: groupValues, tokens: flattenedTokens });
+    flattenTokenGroup({ key: tokenGroup, object: groupValues, tokens: flattenedTokens });
   });
 
   return flattenedTokens;
@@ -89,9 +86,9 @@ function isToken(obj: TokenOrTokenGroup): obj is Token {
 }
 
 /**
- * Helper function used for flattening token files
+ * Helper function that flattens token groups in an object recursively
  */
-function traverseCollection({
+function flattenTokenGroup({
   key,
   object,
   tokens,
@@ -110,7 +107,7 @@ function traverseCollection({
   } else {
     Object.entries<TokenOrTokenGroup>(object).forEach(([key2, object2]) => {
       if (key2.charAt(0) !== '$' && typeof object2 === 'object') {
-        traverseCollection({
+        flattenTokenGroup({
           key: `${key}.${key2}`,
           object: object2,
           tokens,
@@ -120,7 +117,7 @@ function traverseCollection({
   }
 }
 
-function isValidTokenFileName(fileName: string) {
+export function isValidTokenFileName(fileName: string) {
   const fileNameParts = fileName.split('.');
   if (fileNameParts.length < 3 || fileNameParts[2]?.toLowerCase() !== 'json') {
     return false;
@@ -136,45 +133,6 @@ export function collectionAndModeFromFileName(fileName: string) {
   }
   const [collectionName, modeName] = fileName.split('.');
   return { collectionName, modeName };
-}
-
-/**
- * Reads all the tokens in a given tokens directory into an object keyed by the filename.
- * Each file object contains a flattened dictionary of all tokens in that file, keyed by
- * a flattened string representing the key of that token within the hierarchy of its
- * parent token groups. See the doc comment for the `TokensFile` type for more details.
- */
-export function readTokenFiles(tokensDir: string): FlattenedTokensByFile {
-  const files = fs
-    .readdirSync(tokensDir)
-    .filter(isValidTokenFileName)
-    .map((fileName: string) => `${tokensDir}/${fileName}`);
-
-  const tokensJsonByFile: FlattenedTokensByFile = {};
-
-  const seenCollectionsAndModes = new Set<string>();
-
-  files.forEach((file) => {
-    const baseFileName = path.basename(file);
-    const { collectionName, modeName } = collectionAndModeFromFileName(baseFileName);
-
-    if (seenCollectionsAndModes.has(`${collectionName}.${modeName}`)) {
-      throw new Error(`Duplicate collection and mode in file: ${file}`);
-    }
-
-    seenCollectionsAndModes.add(`${collectionName}.${modeName}`);
-
-    const fileContent = fs.readFileSync(file, { encoding: 'utf-8' });
-
-    if (!fileContent) {
-      throw new Error(`Invalid tokens file: ${file}. File is empty.`);
-    }
-    const tokensFile: TokensFile = JSON.parse(fileContent);
-
-    tokensJsonByFile[baseFileName] = flattenTokensFile(tokensFile);
-  });
-
-  return tokensJsonByFile;
 }
 
 const ALIAS_REGEX = /{(.*)}/;
